+++ hahomematic/platforms/custom/cover.py
@@ -28,13 +28,12 @@ from hahomematic.platforms.custom.const import (
 )
 from hahomematic.platforms.custom.entity import CustomEntity
 from hahomematic.platforms.custom.support import CustomConfig, ExtendedConfig
-from hahomematic.platforms.decorators import bind_collector
-from hahomematic.platforms.entity import CallParameterCollector
+from hahomematic.platforms.decorators import value_property
+from hahomematic.platforms.entity import CallParameterCollector, bind_collector
 from hahomematic.platforms.generic.action import HmAction
 from hahomematic.platforms.generic.number import HmFloat
 from hahomematic.platforms.generic.select import HmSelect
 from hahomematic.platforms.generic.sensor import HmSensor
-from hahomematic.platforms.support import value_property
 
 _LOGGER = logging.getLogger("hahomematic.platform")