+++ hahomematic/platforms/custom/switch.py
@@ -19,12 +19,12 @@ from hahomematic.platforms.custom.const import (
 )
 from hahomematic.platforms.custom.entity import CustomEntity
 from hahomematic.platforms.custom.support import CustomConfig, ExtendedConfig
-from hahomematic.platforms.decorators import bind_collector
-from hahomematic.platforms.entity import CallParameterCollector
+from hahomematic.platforms.decorators import value_property
+from hahomematic.platforms.entity import CallParameterCollector, bind_collector
 from hahomematic.platforms.generic.action import HmAction
 from hahomematic.platforms.generic.binary_sensor import HmBinarySensor
 from hahomematic.platforms.generic.switch import HmSwitch
-from hahomematic.platforms.support import OnTimeMixin, value_property
+from hahomematic.platforms.support import OnTimeMixin
 
 _LOGGER = logging.getLogger("hahomematic.platform")