+++ hahomematic/platforms/hub/binary_sensor.py
@@ -6,8 +6,8 @@ See https://www.home-assistant.io/integrations/binary_sensor/.
 from __future__ import annotations
 
 from hahomematic.const import HmPlatform
+from hahomematic.platforms.decorators import value_property
 from hahomematic.platforms.hub.entity import GenericSystemVariable
-from hahomematic.platforms.support import value_property
 
 
 class HmSysvarBinarySensor(GenericSystemVariable):